backend/requirements.txt
fastapi>=0.105.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.27.0
aiosqlite>=0.19.0
httpx>=0.24.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
